chore(ourlogs): Start moving the timeseries events_stats logic to a general place
